Bundle LLVM 21 libraries for shiboken6 runtime support
Fixes #33

The shiboken6 generator was failing at runtime unable to load
libclang 21 shared libraries.

This occurred because shiboken6 was built against LLVM 21 from the SDK
extension (org.freedesktop.Sdk.Extension.llvm21), but these libraries
are not available in the runtime environment.

In this patch, we bundle libLLVM.so.21.1, libclang-cpp.so.21.1, and
libclang.so.21.1 from the LLVM 21 SDK extension into the BaseApp,
along with appropriate symlinks to fix this issues.